refactor: split navigation and ticket components into separate files
